+++ src/main/java/com/educagames/api/model/dto/auth/LoginRequestDTO.java
@@ -1,7 +1,5 @@
 package com.educagames.api.model.dto.auth;
 
-import java.io.Serializable;
-
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
@@ -9,14 +7,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 /**
  * DTO para dados de login.
  */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class LoginRequestDTO implements Serializable {
+public class LoginRequestDTO{
     @Schema(description = "Endereço de e-mail do usuário.", example = "usuario@email.com", requiredMode = Schema.RequiredMode.REQUIRED)
     @Email(message = "Email inválido")
     @NotBlank(message = "O email é obrigatório")